### 2021-07-12 #3487 @AlbertWeichselbraun
Dear all,

Michael Reiss (m.reiss@ikmz.uzh.ch) who is a research and teaching associate at the University of Zurich's Department of Communication and Media Research (IKMZ) might be a good fit.
(he was the first person to request information on how to cite Inscriptis in his work).

From the JOSS reviewer list, reviewers with good Python skills and an interest in Computational Analytics, Natural Language Processing and/or Social Sciences should be a good choice as well. I have located four additional potential reviewers based on these criteria and their position in the reviewer list (preferring reviewers at the bottom of the list):
- ajoer (a.jorgensen@uva.nl)
- hugo.sanjurjo@deusto.es
- alexhanna (alex.hanna@gmail.com)
- mbod (mbod@asc.upenn.edu)

Cheers,
  Albert :)
