### 2021-07-19 5f5af2d
add: link to an article on text extraction from HTML.
