### 2021-08-02 #3557 @whedon
[REVIEW]: Inscriptis - A Python-based HTML to text conversion library optimized for knowledge extraction from the Web
**Submitting author:** @AlbertWeichselbraun (<a href="http://orcid.org/0000-0001-6399-045X">Albert Weichselbraun</a>)
**Repository:** <a href="https://github.com/weblyzard/inscriptis/" target ="_blank">https://github.com/weblyzard/inscriptis/</a>
**Version:** v2.0.0
