### 2021-08-30 #3557 @reality
If a task is not fulfilled completely, should I write the problem here?

1. While the GitHub repository itself does not Apache 2 licence, the repository does not have a LICENSE file in the root.
2. The documentation itself doesn't include a 'statement of need' in the sense that, in its intro, while it mentions it is layout-aware, it doesn't mention particularly that this will aid in knowledge extraction and data science conducted upon web data.
2. Statement of need should include comparison to other HTML parsing and extraction libraries such as cheerio and beautifulsoup. I note that Lynx is used an example of a text-only HTML parser that is not layout aware, but it's is not really a HTML extraction tool. Like you say, many NLP/CS/Biomedical tasks use data from the Web - what software have they been using?
3. Optional: The documentation is overall very good, but the example feels a bit short of what inscriptis can achieve practically. Something like extracting figures or names from a Wikipedia article would really illustrate the usefulness of the software, I think.
4. Given it's on GitHub, users can easily gain support, but there are no contribution guidelines (usually found in the form of CONTRIBUTORS.md)
5. Other software should be cited, e.g. Lynx
