### 2021-09-01 a85aa8c
add: contribution guidelines
These guidelines have been based on the guidelines used by the atom
and gnome-todo projects and been adapted to inscripits.
