+++ docs/installation.rst
@@ -1,19 +1,0 @@
-============
-Installation
-============
-
-Inscriptis requires Python 3.5+ and depends upon the lxml and requests libraries.
-
-At the command line::
-
-    $ pip install inscriptis
-
-Or, if you don't have pip installed::
-
-    $ easy_install inscriptis
-
-If you want to install from the latest sources, you can do::
-
-    $ git clone https://github.com/weblyzard/inscriptis.git
-    $ cd inscriptis
-    $ python setup.py install
