+++ docs/paper/paper.bib
@@ -282,3 +282,156 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	pages = {146--162},
 	file = {Harris_1954_Distributional Structure.pdf:/home/albert/Zotero/storage/59ZE8S3N/Harris_1954_Distributional Structure.pdf:application/pdf;Snapshot:/home/albert/Zotero/storage/HTSHI7XH/00437956.1954.html:text/html},
 }
+
+@misc{html2text,
+    author = {Aaron Swartz},
+    title = {html2text - a Python script that converts a page of HTML into clean, easy-to-read plain ASCII text},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/Alir3z4/html2text},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{justext,
+    author = {Michal Belica},
+    title = {jusText - Heuristic based boilerplate removal tool},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/miso-belica/jusText},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{selenium,
+    author = {Jason Huggins and Paul Gross and Jie Tina Wang},
+    title = {jusText - Heuristic based boilerplate removal tool},
+    year = {2021},
+    publisher = {The Selenium Project},
+    url = {https://www.selenium.dev},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{lynx,
+    author = {Thomas E. Dickey},
+    title = {Lynx - The Text Web-Browser},
+    year = {2021},
+    publisher = {Lynx Home Page},
+    url = {https://lynx.invisible-island.net},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{doccano,
+    author={
+            Hiroki Nakayama and
+            Takahiro Kubo and
+            Junya Kamura and
+            Yasufumi Taniguchi and
+            Xu Liang},
+    title={{doccano}: Text Annotation Tool for Human},
+    year = {2018},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/doccano/doccano},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{beautifulsoup,
+    author = {Leonard Richardson},
+    title = {Beautiful Soup - a library that makes it easy to scrape information from web pages},
+    year = {2021},
+    publisher = {Python Package Index},
+    journal = {PyPI repository},
+    url = {https://pypi.org/project/beautifulsoup4/},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{lxml,
+    author = {Stefan Behnel and
+        Martijn Faassen and
+            Ian Bicking and
+            Holger Joukl and
+            Simon Sapin and 
+            Marc-Antoine Parent and
+            Olivier Grisel and 
+            Kasimier Buchcik and
+            Florian Wagner and
+            Emil Kroymann and
+            Paul Everitt and 
+            Victor Ng and
+            Robert Kern and
+            Andreas Pakulat and
+            David Sankel and
+            Marcin Kasperski and
+            Sidnei da Silva and
+            Pascal Oberndörfer},
+    title = {lxml - Processing XML and HTML with Python},
+    year = {2021},
+    publisher = {lxml Project},
+    url = {https://lxml.de/},
+    note = {Accessed: 2021-09-02}
+}
+
+%% auhors are based on https://yarnpkg.com/package/cheerio
+@misc{cheerio,
+    author = {Matthew Mueller and Felix Böhm and Jugglin Mike and David Chambers},
[…]
