+++ docs/paper/paper.md
@@ -22,8 +22,8 @@ bibliography: paper.bib
  
 ``Inscriptis`` provides a library, command line client and Web service for converting HTML to plain text.
  
-Its development has been triggered by the need to obtain accurate text representations for knowledge extraction tasks that preserve the spatial alignment of text without drawing upon heavyweight, browser-based solutions such as [Selenium](https://www.selenium.dev/).
-In contrast to existing software packages such as [HTML2text](https://github.com/Alir3z4/html2text/), [jusText](https://github.com/miso-belica/jusText/) and [Lynx](https://lynx.invisible-island.net/), ``Inscriptis``
+Its development has been triggered by the need to obtain accurate text representations for knowledge extraction tasks that preserve the spatial alignment of text without drawing upon heavyweight, browser-based solutions such as Selenium[@selenium].
+In contrast to existing software packages such as HTML2text [@html2text], jusText [@justext] and Lynx[@lynx], ``Inscriptis``
  
 1. provides a layout-aware conversion of HTML that more closely resembles the rendering obtained from standard Web browsers and, therefore, better preserves the spatial arrangement of text elements. ``Inscriptis`` excels in terms of conversion quality, since it correctly converts complex HTML constructs such as nested tables and also interprets a subset of HTML (e.g., `align`, `valign`) and CSS (e.g., `display`, `white-space`, `margin-top`, `vertical-align`, etc.) attributes that determine the text alignment.
 2. supports annotation rules, i.e., user-provided mappings that allow for annotating the extracted text based on structural and semantic information encoded in HTML tags and attributes used for controlling structure and layout in the original HTML document.
@@ -42,17 +42,23 @@ Most knowledge extraction methods operate on text and, therefore, require an acc
 Despite this need from within the research community, many standard HTML to text conversion techniques are not layout aware, yielding text representations that fail to preserve the text's spatial properties, as illustrated below:
  
 ![Text representation of a table from Wikipedia computed by ``Inscriptis`` (left) and lynx (right). Lynx fails to correctly interpret the table and, therefore, does not properly align the temperature values.](images/inscriptis-vs-lynx.png)
- 
-``Inscriptis`` is not only able to correctly render such pages but also offers the option to preserve parts of the original HTML document's semantics (e.g., information on headings, emphasized text, tables, etc.) by complementing the extracted text with annotations obtained from the document. \autoref{fig:annotations} provides an example of annotations extracted from a Wikipedia page. These annotations can be useful for
+
+Consequently, even popular resources which have been extensively used in literature, suffer from such shortcomings. The Common Crawl dataset which contains billions of Web documents and has been among the datasets used for training transformer models such as ..., for instance, provides text representations which have been generated with a custom converter HTML converters that does not consider layout information.
+
+Even worse, some tutorials suggest the use of non-specialized software libraries such as Beautifulsoup [@beautifulsoup], lxml [@lxml] and Cheerio [@cheerio] for text extraction although these libraries have been designed with a different use case in mind and, therefore, do not interpret the semantics of HTML. Although efficient, if used for extracting text only content, this approach even yields artifacts (e.g., concatenated words) with simple HTML constructs such as lists rather than a properly formatted output. Some libraries such as Cheerio even warn their users, by explicitely stating that they are not well-suited for emulating Web browsers.
+
[…]
