+++ docs/paper/paper.bib
@@ -283,6 +283,17 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	file = {Harris_1954_Distributional Structure.pdf:/home/albert/Zotero/storage/59ZE8S3N/Harris_1954_Distributional Structure.pdf:application/pdf;Snapshot:/home/albert/Zotero/storage/HTSHI7XH/00437956.1954.html:text/html},
 }
 
+@misc{ia-commons,
+    author = {Ilya Kreymer and Sebastian Nagel and Andy Jackson and Noah Levitt},
+    title = {IIPC Web Archive Commons - Utility code for OpenWayBack and other projects.},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/commoncrawl/ia-web-commons},
+    note = {Accessed: 2021-09-02}
+}
+
+
 @misc{html2text,
     author = {Aaron Swartz},
     title = {html2text - a Python script that converts a page of HTML into clean, easy-to-read plain ASCII text},
@@ -435,3 +446,69 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	file = {lang2012-textSweeper.pdf:/home/albert/Zotero/storage/BKDTK3SJ/lang2012-textSweeper.pdf:application/pdf},
     url = {https://aisel.aisnet.org/confirm2012/17/},
 }
+
+
+@article{caswell_quality_2021,
+	title = {Quality at a {Glance}: {An} {Audit} of {Web}-{Crawled} {Multilingual} {Datasets}},
+	shorttitle = {Quality at a {Glance}},
+	url = {http://arxiv.org/abs/2103.12028},
+	abstract = {With the success of large-scale pre-training and multilingual modeling in Natural Language Processing (NLP), recent years have seen a proliferation of large, web-mined text datasets covering hundreds of languages. However, to date there has been no systematic analysis of the quality of these publicly available datasets, or whether the datasets actually contain content in the languages they claim to represent. In this work, we manually audit the quality of 205 language-specific corpora released with five major public datasets (CCAligned, ParaCrawl, WikiMatrix, OSCAR, mC4), and audit the correctness of language codes in a sixth (JW300). We find that lower-resource corpora have systematic issues: at least 15 corpora are completely erroneous, and a significant fraction contains less than 50\% sentences of acceptable quality. Similarly, we find 82 corpora that are mislabeled or use nonstandard/ambiguous language codes. We demonstrate that these issues are easy to detect even for non-speakers of the languages in question, and supplement the human judgements with automatic analyses. Inspired by our analysis, we recommend techniques to evaluate and improve multilingual corpora and discuss the risks that come with low-quality data releases.},
+	urldate = {2021-09-03},
+	journal = {arXiv:2103.12028 [cs]},
[…]
