+++ docs/paper/paper.md
@@ -41,15 +41,15 @@ Most knowledge extraction methods operate on text and, therefore, require an acc
  
 Despite this need from within the research community, many standard HTML to text conversion techniques are not layout aware, yielding text representations that fail to preserve the text's spatial properties, as illustrated below:
  
-![Text representation of a table from Wikipedia computed by ``Inscriptis`` (left) and lynx (right). Lynx fails to correctly interpret the table and, therefore, does not properly align the temperature values.](images/inscriptis-vs-lynx.png)
+![Text representation of a table from Wikipedia computed by ``Inscriptis`` (left) and Lynx (right). Lynx fails to correctly interpret the table and, therefore, does not properly align the temperature values.](images/inscriptis-vs-lynx.png)
 
-Consequently, even popular resources which have been extensively used in literature, suffer from such shortcomings. The Common Crawl dataset which contains billions of Web documents and has been among the datasets used for training transformer models such as ..., for instance, provides text representations which have been generated with a custom converter HTML converters that does not consider layout information.
+Consequently, even popular resources extensively used in literature suffer from such shortcomings. The text representations provided with the Common Crawl corpus^[https://commoncrawl.org/], for instance, have been generated with a custom utility [@ia-commons] which at the time of writing did not consider any layout information. Datasets such as CCAligned [@el-kishky_ccaligned_2020], multilingual C4 which has been used for training the mT5 language model [@xue_mt5_2021], and OSCAR [@suarez_asynchronous_2019] are based on subsets of the Common Crawl corpus [@caswell_quality_2021].
 
-Even worse, some tutorials suggest the use of non-specialized software libraries such as Beautifulsoup [@beautifulsoup], lxml [@lxml] and Cheerio [@cheerio] for text extraction although these libraries have been designed with a different use case in mind and, therefore, do not interpret the semantics of HTML. Although efficient, if used for extracting text only content, this approach even yields artifacts (e.g., concatenated words) with simple HTML constructs such as lists rather than a properly formatted output. Some libraries such as Cheerio even warn their users, by explicitely stating that they are not well-suited for emulating Web browsers.
+Even worse, some tutorials suggest the use of non-specialized software libraries such as Beautiful Soup [@beautifulsoup], lxml [@lxml] and Cheerio [@cheerio] for converting HTML. Since these libraries have been designed with a different use case in mind, they are well-suited for scraping text only content. Once they encounter HTML constructs such as lists and tables, these libraries are likely to return artifacts (e.g., concatenated words), since they do not interpret HTML semantics. The creators of the Cheerio library even warn their users, by explicitly stating that it is not well-suited for emulating Web browsers.
 
-Specialized conversion tools such as HTML2Text perform considerably better but they often fail for more complex pages. Researchers sometimes even use text-based Web browsers such as Lynx to obtain more accurate representations of HTML pages. These tools are complemented by content extraction software such as jusText [@justext], dragnet [@peters_content_2013], TextSweeper [@lang_textsweeper_2012] and boilerpy3 [@boilerpy3] which do not consider the page layout but rather aim at extracting the relevant content only, and approaches that are optimized for certain kinds of Web pages like Harvest [@weichselbraun_harvest_2020] for Web forums.
[…]
