+++ docs/paper/paper.bib
@@ -102,6 +102,7 @@ Gui et al. - public dataset based on SINA city news.
 	month = jan,
 	year = {2016},
 	note = {Accepted 17 August 2015},
+    doi = {10.1109/hicss.2016.133},
 	keywords = {climate change, keyword analysis, Opinion mining, opinion target extraction, sentiment analysis},
 }
 
@@ -243,6 +244,7 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	note = {arXiv: 2106.00641},
 	keywords = {Computer Science - Computation and Language},
 	annote = {Comment: Accepted by ACL 2021 (Main track)},
+    doi = {10.18653/v1/2021.acl-long.558},
 	file = {arXiv.org Snapshot:/home/albert/Zotero/storage/4LF375ZK/2106.html:text/html;Fu et al_2021_SpanNER.pdf:/home/albert/Zotero/storage/3IEZ8VGV/Fu et al_2021_SpanNER.pdf:application/pdf},
 }
 
@@ -454,7 +456,6 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	url = {http://arxiv.org/abs/2103.12028},
 	abstract = {With the success of large-scale pre-training and multilingual modeling in Natural Language Processing (NLP), recent years have seen a proliferation of large, web-mined text datasets covering hundreds of languages. However, to date there has been no systematic analysis of the quality of these publicly available datasets, or whether the datasets actually contain content in the languages they claim to represent. In this work, we manually audit the quality of 205 language-specific corpora released with five major public datasets (CCAligned, ParaCrawl, WikiMatrix, OSCAR, mC4), and audit the correctness of language codes in a sixth (JW300). We find that lower-resource corpora have systematic issues: at least 15 corpora are completely erroneous, and a significant fraction contains less than 50\% sentences of acceptable quality. Similarly, we find 82 corpora that are mislabeled or use nonstandard/ambiguous language codes. We demonstrate that these issues are easy to detect even for non-speakers of the languages in question, and supplement the human judgements with automatic analyses. Inspired by our analysis, we recommend techniques to evaluate and improve multilingual corpora and discuss the risks that come with low-quality data releases.},
 	urldate = {2021-09-03},
-	journal = {arXiv:2103.12028 [cs]},
[…]
