+++ docs/paper/paper.md
@@ -22,8 +22,8 @@ bibliography: paper.bib
  
 ``Inscriptis`` provides a library, command line client and Web service for converting HTML to plain text.
  
-Its development has been triggered by the need to obtain accurate text representations for knowledge extraction tasks that preserve the spatial alignment of text without drawing upon heavyweight, browser-based solutions such as Selenium[@selenium].
-In contrast to existing software packages such as HTML2text [@html2text], jusText [@justext] and Lynx[@lynx], ``Inscriptis``
+Its development has been triggered by the need to obtain accurate text representations for knowledge extraction tasks that preserve the spatial alignment of text without drawing upon heavyweight, browser-based solutions such as Selenium [@selenium].
+In contrast to existing software packages such as HTML2text [@html2text], jusText [@justext] and Lynx [@lynx], ``Inscriptis``
  
 1. provides a layout-aware conversion of HTML that more closely resembles the rendering obtained from standard Web browsers and, therefore, better preserves the spatial arrangement of text elements. ``Inscriptis`` excels in terms of conversion quality, since it correctly converts complex HTML constructs such as nested tables and also interprets a subset of HTML (e.g., `align`, `valign`) and CSS (e.g., `display`, `white-space`, `margin-top`, `vertical-align`, etc.) attributes that determine the text alignment.
 2. supports annotation rules, i.e., user-provided mappings that allow for annotating the extracted text based on structural and semantic information encoded in HTML tags and attributes used for controlling structure and layout in the original HTML document.
@@ -45,7 +45,7 @@ Despite this need from within the research community, many standard HTML to text
 
 Consequently, even popular resources extensively used in literature suffer from such shortcomings. The text representations provided with the Common Crawl corpus^[https://commoncrawl.org/], for instance, have been generated with a custom utility [@ia-commons] which at the time of writing did not consider any layout information. Datasets such as CCAligned [@el-kishky_ccaligned_2020], multilingual C4 which has been used for training the mT5 language model [@xue_mt5_2021], and OSCAR [@suarez_asynchronous_2019] are based on subsets of the Common Crawl corpus [@caswell_quality_2021].
 
-Even worse, some tutorials suggest the use of non-specialized software libraries such as Beautiful Soup [@beautifulsoup], lxml [@lxml] and Cheerio [@cheerio] for converting HTML. Since these libraries have been designed with a different use case in mind, they are well-suited for scraping text only content. Once they encounter HTML constructs such as lists and tables, these libraries are likely to return artifacts (e.g., concatenated words), since they do not interpret HTML semantics. The creators of the Cheerio library even warn their users, by explicitly stating that it is not well-suited for emulating Web browsers.
+Even worse, some tutorials suggest the use of software libraries such as Beautiful Soup [@beautifulsoup], lxml [@lxml] and Cheerio [@cheerio] for converting HTML. Since these libraries have been designed with a different use case in mind, they are only well-suited for scraping textual content. Once they encounter HTML constructs such as lists and tables, these libraries are likely to return artifacts (e.g., concatenated words), since they do not interpret HTML semantics. The creators of the Cheerio library even warn their users, by explicitly stating that it is not well-suited for emulating Web browsers.
 
 Specialized conversion tools such as HTML2Text perform considerably better but often fail for more complex Web pages. Researchers sometimes even draw upon text-based Web browsers such as Lynx to obtain more accurate representations of HTML pages. These tools are complemented by content extraction software such as jusText [@justext], dragnet [@peters_content_2013], TextSweeper [@lang_textsweeper_2012] and boilerpy3 [@boilerpy3] which do not consider the page layout but rather aim at extracting the relevant content only, and approaches that are optimized for certain kinds of Web pages like Harvest [@weichselbraun_harvest_2020] for Web forums.
 
