### 2021-09-03 a3d6c7d
chg: do not consider conf.py for codefactor.io
