+++ docs/conf.py
@@ -54,6 +54,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'inscriptis'
+
 copyright = __copyright__
 author = __author__
 
