### 2021-09-03 31a0f44
chg: do not consider conf.py for codefactor.io
