### 2021-09-03 e23f79a
Fix/joss review 2021 08 30 (#37)
* fix: use LICENSE rather than COPYING for the license file.

* add: contribution guidelines

These guidelines have been based on the guidelines used by the atom
and gnome-todo projects and been adapted to inscripits.

* chg: refined documentation.

* add: more examples for annotation profiles.

* chg: improved examples.

* fix: documentation on annotation rules.

* wip: improved statement of need.

* add: improved statement of need based on the reviewer comments

* fix: grammar, spelling and linking erros.

* chg: extended guidelines.

* fix: spacing and phras
[…]
