+++ docs/paper/paper.bib
@@ -102,6 +102,7 @@ Gui et al. - public dataset based on SINA city news.
 	month = jan,
 	year = {2016},
 	note = {Accepted 17 August 2015},
+    doi = {10.1109/hicss.2016.133},
 	keywords = {climate change, keyword analysis, Opinion mining, opinion target extraction, sentiment analysis},
 }
 
@@ -243,6 +244,7 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	note = {arXiv: 2106.00641},
 	keywords = {Computer Science - Computation and Language},
 	annote = {Comment: Accepted by ACL 2021 (Main track)},
+    doi = {10.18653/v1/2021.acl-long.558},
 	file = {arXiv.org Snapshot:/home/albert/Zotero/storage/4LF375ZK/2106.html:text/html;Fu et al_2021_SpanNER.pdf:/home/albert/Zotero/storage/3IEZ8VGV/Fu et al_2021_SpanNER.pdf:application/pdf},
 }
 
@@ -282,3 +284,232 @@ similar to WISDOM - identify businesses in financial news and propagate the impa
 	pages = {146--162},
 	file = {Harris_1954_Distributional Structure.pdf:/home/albert/Zotero/storage/59ZE8S3N/Harris_1954_Distributional Structure.pdf:application/pdf;Snapshot:/home/albert/Zotero/storage/HTSHI7XH/00437956.1954.html:text/html},
 }
+
+@misc{ia-commons,
+    author = {Ilya Kreymer and Sebastian Nagel and Andy Jackson and Noah Levitt},
+    title = {IIPC Web Archive Commons - Utility code for OpenWayBack and other projects.},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/commoncrawl/ia-web-commons},
+    note = {Accessed: 2021-09-02}
+}
+
+
+@misc{html2text,
+    author = {Aaron Swartz},
+    title = {html2text - a Python script that converts a page of HTML into clean, easy-to-read plain ASCII text},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/Alir3z4/html2text},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{justext,
+    author = {Michal Belica},
+    title = {jusText - Heuristic based boilerplate removal tool},
+    year = {2021},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/miso-belica/jusText},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{selenium,
+    author = {Jason Huggins and Paul Gross and Jie Tina Wang},
+    title = {jusText - Heuristic based boilerplate removal tool},
+    year = {2021},
+    publisher = {The Selenium Project},
+    url = {https://www.selenium.dev},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{lynx,
+    author = {Thomas E. Dickey},
+    title = {Lynx - The Text Web-Browser},
+    year = {2021},
+    publisher = {Lynx Home Page},
+    url = {https://lynx.invisible-island.net},
+    note = {Accessed: 2021-09-02}
+}
+
+@misc{doccano,
+    author={
+            Hiroki Nakayama and
+            Takahiro Kubo and
+            Junya Kamura and
+            Yasufumi Taniguchi and
+            Xu Liang},
+    title={{doccano}: Text Annotation Tool for Human},
+    year = {2018},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/doccano/doccano},
+    note = {Accessed: 2021-09-02}
+}
+
[…]
