### 2021-09-03 #3557 @AlbertWeichselbraun
Dear @reality, dear @sbenthall

Thank you for this valuable feedback. Please find below a detailed description of how I addressed the suggestions outlined in the review. All the listed improvements have been merged into the project's master branch some minutes ago.

Best regards,
  Albert

--

**Issue 1:** For historical reasons, the license has been in the COPYING file. I have renamed the file to LICENSE to address this comment.

**Issue 2:** Based on your suggestion, the project's documentation has been extended to include a statement of need. In addition, the introduction paragraph now also mentions that Inscriptis is particularly well suited for knowledge extraction and data science tasks.

**Issue 3:** Thank you for pointing this out. I have added a comparison with related software to the paper's "statement of need" section, that considers Beautiful Soup, lxml, Cheerio, HTML2Text, Lynx, jusText, TextSweeper, boilerpy3 and HARVEST. These libraries have been discussed based on the following classification:

   1. libraries that are focused on other use cases, therefore, do not interpret HTML semantics, and are likely to produce serious conversion errors (Beautiful Soup, lxml, Cheerio);
   2. conversion software that provides reasonable good text representations of HTML (HTML2Text, Lynx); and
   3. content-extraction tools that focus on specialized tasks such as boilerplate removal and forum extraction (jusText, TextSweeper, boilerpy3, HARVEST).

  In addition, the statement of need now emphasizes the importance of Inscriptis by pointing out that even the popular Common Crawl corpus does not use a layout-aware HTML to text conversion approach, which in turn affects derived datasets such as CCAligned, multilingual C4 and OSCAR, and techniques that build upon them (e.g., the mT5 language model).

**Issue 4:** Based on your feedback, I have added examples of Inscriptis' annotation support to the documentation that cover the following four use cases:

   - Wikipedia tables and table metadata
   - References to entities, missing entities and citations from Wikipedia
   - Posts and post metadata from the XDA developer forum
   - Code and metadata from Stackoverflow pages

The prior section on Inscriptis' annotation support still uses the simple "toy" example, since it provides an easy-to-understand use case of how annotations are supposed to work.

**Issue 5:** Thank you for this suggestion. I have added a `CONTRIBUTING.md` file to the repository, which has also been integrated into the Inscriptis documentation.

**Issue 6:** Based on your feedback, I have replaced the URLs to other software with citations.
