### 2021-09-03 101ae76
Adjust annotation example instructions
Instructions are written as if the user has the repository checked out, but preceding instructions use pip install or easy_install
