### 2021-09-03 weblyzard/inscriptis#39 @rlskoeser
tag preprocessor example in readme results in an error
the readme includes this example usage under the annotation preprocessors section:
```
$ inscript.py https://www.fhgr.ch \
        -r ./examples/annotation-profile.json \
        -p tag
```

When I try running it, I get this error:
```
ModuleNotFoundError: No module named 'inscriptis.annotation.output.tag'
```

It's not clear from the example whether it's meant to be a placeholder command or a real one (I think the first?); maybe just revise it to use an actual supported preprocessor to avoid confusion
