### 2021-09-03 weblyzard/inscriptis#40 @rlskoeser
web service instructions don't work as written
As far as I can tell the web service only works if you check out the git repository, not from a pip install.  You might think about including it in the python package, but I don't know how important that is.

I think it would be sufficient to amend the instructions to add directions to check out the repository, and then correct the path to `web-service.py` which is now under the `scripts` directory.
