### 2021-09-03 #3557 @rlskoeser
Apologies for the slow review, have been a bit overwhelmed by other commitments.

I worked through the instructions and opened a PR and couple of issues to correct and point out some minor problems I found in the directions; I think they should all be pretty trivial to address.

Thanks to the helpful review from @reality and your work to address them it seems like this is in good shape.
