### 2021-09-06 7fbb561
fix: web service instructions don't work (#40)
- moved web service into the inscriptis package.
- adjusted documentation and Dockerfile accordingly.
