+++ Dockerfile
@@ -9,7 +9,6 @@ RUN pip install --no-cache-dir -r requirements.txt
 RUN pip install --no-cache-dir Flask
 RUN python setup.py install
 
-# RUN export FLASK_APP="web-service.py"
-# CMD ["python3", "-m", "flask", "run"]
+RUN export FLASK_APP="inscriptis.service.web"
+CMD ["python3", "-m", "flask", "run"]
 EXPOSE 5000
-CMD ["python3", "scripts/web-service.py"]
