### 2021-09-06 weblyzard/inscriptis#40 @AlbertWeichselbraun
Thank you for pointing this out. I have moved the web service into the Inscriptis package and adjusted the documentation accordingly.
