+++ Dockerfile
@@ -9,6 +9,6 @@ RUN pip install --no-cache-dir -r requirements.txt
 RUN pip install --no-cache-dir Flask
 RUN python setup.py install
 
-RUN export FLASK_APP="inscriptis.service.web"
+ENV FLASK_APP="inscriptis.service.web"
 CMD ["python3", "-m", "flask", "run"]
 EXPOSE 5000
