### 2021-09-06 d235a72
chg: consolidated run statements in Dockerfile.
