### 2021-09-06 #3557 @AlbertWeichselbraun
Dear @rlskoeser, dear @sbenthall,

Thank you for your valuable feedback, and especially for spotting the reported documentation issues. Please find below a detailed description of how I addressed the outlined suggestions:

1. I have merged the [documentation fixes](https://github.com/weblyzard/inscriptis/pull/38) with the master branch.
2. The documentation bug reported in [#39](https://github.com/weblyzard/inscriptis/issues/39) has been resolved.
3. I have added the Inscriptis web service to the Python package and updated the documentation accordingly, resolving [#40](https://github.com/weblyzard/inscriptis/issues/40).

Best regards,
  Albert

--
