### 2021-09-27 #3557 @AlbertWeichselbraun
Dear @reality and @sbenthall,

@reality's [review](https://github.com/openjournals/joss-reviews/issues/3557#issuecomment-908259651) has triggered the improvements summarized in my [comment from 3 September 2021](https://github.com/openjournals/joss-reviews/issues/3557#issuecomment-912503256).

To the best of my knowledge this should address all the issues raised in the review and the missing tasks in the review checklist. Please let me know if there is still any item that has not been completely fulfilled and/or needs improvement.

Cheers,
   Albert :-)
