### 2021-10-11 27865db
add: docstring to webservice __init__ file.
