+++ src/inscriptis/service/__init__.py
@@ -1,0 +1,1 @@
+"""The Inscriptis Web service."""
